refactor: extract shell and VMM into separate modules with enhanced shell features (#340)
* refactor: extract shell and vmm into separate modules

* fix: update paths for axvisor-vmm and axvisor-shell to relative references

* feat: Implement tab completion and command history management

* refactor: integrate shell and vmm modules directly into kernel
