+++ kernel/src/shell/parser/mod.rs
@@ -1,0 +1,11 @@
+//! Command parsing module
+//!
+//! Provides functionality to parse and structure commands.
+
+mod history;
+mod node;
+mod parser;
+
+pub use history::{CommandHistory, clear_line_and_redraw};
+pub use node::{CommandNode, FlagDef, OptionDef, ParseError, ParsedCommand};
+pub use parser::CommandParser;
