+++ kernel/src/vmm/vm_list.rs
@@ -1,5 +1,4 @@
 use alloc::{collections::BTreeMap, sync::Arc, vec::Vec};
-
 use spin::Mutex;
 
 pub type VMRef = Arc<axvm::Vm>;
