+++ modules/axfs/src/fs/fatfs.rs
@@ -4,8 +4,8 @@ use core::cell::OnceCell;
 
 use axfs_vfs::{VfsDirEntry, VfsError, VfsNodePerm, VfsResult};
 use axfs_vfs::{VfsNodeAttr, VfsNodeOps, VfsNodeRef, VfsNodeType, VfsOps};
-use spin::Mutex;
 use fatfs::{Dir, File, LossyOemCpConverter, NullTimeProvider, Read, Seek, SeekFrom, Write};
+use spin::Mutex;
 
 use crate::dev::{Disk, Partition};
